[refactor] Layout 폴더 정리
