+++ src/components/Layout/LoginForm/index.tsx
@@ -2,15 +2,15 @@
 
 // Compo
 import Button from "@/components/base/Button";
-import InputWithLabel from "@/components/InputWithLabel";
+import InputWithLabel from "@/components/Composition/InputWithLabel";
 import ShleterSelect from "@/components/base/CustomSelectBox";
 // Utils
 import React, { useState } from "react";
 import { setCookie } from "@/utils/cookie";
 import { useMutation } from "@tanstack/react-query";
-import { login } from "@/utils/api/v1/shelter-admin";
+import { login } from "@/api/v1/shelter-admin";
 //Types
-import type { LoginBodyType } from "@/utils/api/v1/shelter-admin/type";
+import type { LoginBodyType } from "@/api/v1/shelter-admin/type";
 
 const LoginForm = () => {
   const { mutate } = useMutation({
